backend/python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
redis[asyncio]==5.2.0
apscheduler==3.10.4
python-slugify==8.0.4
python-multipart==0.0.12
feedgen==1.0.0
